:hammer: chore: code improvements
